ai-capability-skills-agent-persona/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-capability-skills-agent-persona"
version = "0.1.0"
description = "Standalone microservice for managing AI agent personas, capabilities, and skills — config-driven extensibility for multi-agent workflows"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Aarush" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "watchdog>=4.0.0",
    "jsonschema>=4.22.0",
    "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[project.scripts]
capability-service = "src.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
